requirements.txt
streamlit==1.32.0
torch==2.2.1
torchvision==0.17.1
timm==0.9.16
Pillow==10.2.0
numpy==1.26.4
